Safari-609.1.11.1/LayoutTests/media/video-main-content-allow-expected.txt

RUN(internals.setMediaElementRestrictions(video, "RequireUserGestureForVideoRateChange,OverrideUserGestureRequirementForMainContent"))
EVENT(canplaythrough)
EVENT(playing)
END OF TEST
